+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: `Cards`: Add `PartnerCard` component.
+
 ## [12.14.0] - 2022-10-14  💅
 
 Feature: